requirements.txt
pandas==2.2.0
requests==2.31.0
python-dotenv==1.0.0
numpy==1.26.3